rename traits and functions
